Create task3_client.cpp
The little that I have so far for the task3 client. Since it's a stream socket, the messages have been getting received as a single long message. Search space still needs to be split up between clients and the clients still need to be programmed to conduct the search for the nonce.